Server/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
pydrive
Pillow
